Adjuustment to send model across to files without context
This was an idea i had pertaining to a similar issue.
I couldn't figure out the details before but then I had this idea to just simply cast it at the start if the struct is wrong and it will return an error.

I'm pretty sure this should work in theory. Assuming it does I can use similar logic on 'RunUntillSuccesfull' so that it can go into the common folder.